clean up public token implementation
