+++ src/token/mod.rs
@@ -8,6 +8,7 @@ use std::time::Duration;
 pub use claims::Claims;
 pub use public::SigningKey;
 
+use bytes::{BufMut, Bytes, BytesMut};
 use thiserror::Error;
 use time::OffsetDateTime;
 use tracing::error;
@@ -44,5 +45,42 @@ impl UserToken {
             error!("error signing paseto");
             Error::SigningError
         })
+    }
+}
+
+fn pre_auth_encode(pieces: &[&[u8]]) -> Bytes {
+    let mut capacity = 8;
+    for piece in pieces {
+        capacity += 8 + piece.len();
+    }
+
+    // bitmasks since the Paseto standard requires that the highest bit of
+    // each unsigned integer is unset
+    let mut pae = BytesMut::with_capacity(capacity);
+    pae.put_u64_le(pieces.len() as u64 & 0x7FFFFFFFFFFFFF);
+    for piece in pieces {
+        pae.put_u64_le(piece.len() as u64 & 0x7FFFFFFFFFFFFF);
+        pae.put_slice(piece);
+    }
+
+    pae.freeze()
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    use hex_literal::hex;
+
+    #[test]
+    fn pae() {
+        let one = pre_auth_encode(&[]);
+        assert_eq!(*one, hex!("0000000000000000"));
+
+        let two = pre_auth_encode(&[b""]);
+        assert_eq!(*two, hex!("0100000000000000 0000000000000000"));
+
+        let three = pre_auth_encode(&[b"test"]);
+        assert_eq!(*three, hex!("0100000000000000 0400000000000000 74657374"));
     }
 }
