+++ src/token/public.rs
@@ -1,15 +1,13 @@
 //! Paseto version 4 public tokens.
 
-use super::claims::Claims;
+use crate::token::{claims::Claims, pre_auth_encode};
 
 use base64::{engine::general_purpose::URL_SAFE_NO_PAD, Engine};
-use bytes::{BufMut, Bytes, BytesMut};
-use ed25519::Signature;
 use pkcs8::DecodePrivateKey;
 use signature::Signer;
 use thiserror::Error;
 
-const PUBLIC_HEADER: &str = "v4.public.";
+pub const PUBLIC_HEADER: &str = "v4.public.";
 
 /// Errors while interacting with public tokens.
 #[derive(Error, Clone, Debug)]
@@ -40,20 +38,15 @@ impl SigningKey {
         footer: Option<&[u8]>,
         implicit: Option<&[u8]>,
     ) -> String {
-        let pae = pre_auth_encode(&[
+        let sig = self.0.sign(&pre_auth_encode(&[
             PUBLIC_HEADER.as_bytes(),
             message,
             footer.unwrap_or(&[]),
             implicit.unwrap_or(&[]),
-        ]);
-        let sig = self.0.sign(&pae);
-
-        let mut body = Vec::with_capacity(message.len() + Signature::BYTE_SIZE);
-        body.extend_from_slice(message);
-        body.extend_from_slice(&sig.to_bytes());
+        ]));
 
         let mut token = PUBLIC_HEADER.to_string();
-        token += &URL_SAFE_NO_PAD.encode(&body);
+        token += &URL_SAFE_NO_PAD.encode([message, &sig.to_bytes()].concat());
         if let Some(footer) = footer {
             token += ".";
             token += &URL_SAFE_NO_PAD.encode(footer);
@@ -81,69 +74,58 @@ impl SigningKey {
     }
 }
 
-fn pre_auth_encode(pieces: &[&[u8]]) -> Bytes {
-    let mut capacity = 8;
-    for piece in pieces {
-        capacity += 8 + piece.len();
-    }
-
-    // bitmasks since the Paseto standard requires that the highest bit of
-    // each unsigned integer is unset
-    let mut pae = BytesMut::with_capacity(capacity);
-    pae.put_u64_le(pieces.len() as u64 & 0x7FFFFFFFFFFFFF);
-    for piece in pieces {
-        pae.put_u64_le(piece.len() as u64 & 0x7FFFFFFFFFFFFF);
-        pae.put_slice(piece);
-    }
-
-    pae.freeze()
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
 
     use hex_literal::hex;
 
+    const KEY_PEM: &str = "-----BEGIN PRIVATE KEY-----\nMC4CAQAwBQYDK2VwBCIEILTL+0PfTOIQcn2VPkpxMwf6Gbt9n4UEFDjZ4RuUKjd0\n-----END PRIVATE KEY-----\n";
+
     #[test]
     fn decode_pem() {
-        let secret_key = hex!("DF6604C6CBA2BE10CE89997C100DE0B3EF75CFAD36084651C2C983773B8C0F8F");
-        let pem = "-----BEGIN PRIVATE KEY-----\nMC4CAQAwBQYDK2VwBCIEIN9mBMbLor4QzomZfBAN4LPvdc+tNghGUcLJg3c7jA+P\n-----END PRIVATE KEY-----\n";
+        let key_bytes = hex!("B4CBFB43DF4CE210727D953E4A713307FA19BB7D9F85041438D9E11B942A3774");
 
-        let signing_key = SigningKey::from_pem(pem).expect("decode pem");
-
-        assert_eq!(secret_key, signing_key.0.to_bytes())
-    }
-
-    #[test]
-    fn pae() {
-        let one = pre_auth_encode(&[]);
-        assert_eq!(*one, hex!("0000000000000000"));
-
-        let two = pre_auth_encode(&[b""]);
-        assert_eq!(*two, hex!("0100000000000000 0000000000000000"));
-
-        let three = pre_auth_encode(&[b"test"]);
-        assert_eq!(*three, hex!("0100000000000000 0400000000000000 74657374"));
+        let key = SigningKey::from_pem(KEY_PEM).expect("decode pem");
+        assert_eq!(key.0.to_bytes(), key_bytes);
     }
 
     #[test]
     fn public_sign() {
-        let pem = "-----BEGIN PRIVATE KEY-----\nMC4CAQAwBQYDK2VwBCIEILTL+0PfTOIQcn2VPkpxMwf6Gbt9n4UEFDjZ4RuUKjd0\n-----END PRIVATE KEY-----\n";
-        let signing_key = SigningKey::from_pem(pem).expect("decode pem");
+        let payload =
+            b"{\"data\":\"this is a signed message\",\"exp\":\"2022-01-01T00:00:00+00:00\"}";
+        let expected = "v4.public.eyJkYXRhIjoidGhpcyBpcyBhIHNpZ25lZCBtZXNzYWdlIiwiZXhwIjoiMjAyMi0wMS0wMVQwMDowMDowMCswMDowMCJ9bg_XBBzds8lTZShVlwwKSgeKpLT3yukTw6JUz3W4h_ExsQV-P0V54zemZDcAxFaSeef1QlXEFtkqxT1ciiQEDA";
 
+        let key = SigningKey::from_pem(KEY_PEM).expect("decode pem");
+
+        let token = key.sign_message(payload, None, None);
+        assert_eq!(token, expected);
+    }
+
+    #[test]
+    fn public_sign_with_footer() {
+        let payload =
+            b"{\"data\":\"this is a signed message\",\"exp\":\"2022-01-01T00:00:00+00:00\"}";
+        let footer = b"{\"kid\":\"zVhMiPBP9fRf2snEcT7gFTioeA9COcNy9DfgL1W60haN\"}";
+        let expected = "v4.public.eyJkYXRhIjoidGhpcyBpcyBhIHNpZ25lZCBtZXNzYWdlIiwiZXhwIjoiMjAyMi0wMS0wMVQwMDowMDowMCswMDowMCJ9v3Jt8mx_TdM2ceTGoqwrh4yDFn0XsHvvV_D0DtwQxVrJEBMl0F2caAdgnpKlt4p7xBnx1HcO-SPo8FPp214HDw.eyJraWQiOiJ6VmhNaVBCUDlmUmYyc25FY1Q3Z0ZUaW9lQTlDT2NOeTlEZmdMMVc2MGhhTiJ9";
+
+        let key = SigningKey::from_pem(KEY_PEM).expect("decode pem");
+
+        let token = key.sign_message(payload, Some(footer), None);
+        assert_eq!(token, expected);
+    }
+
+    #[test]
+    fn public_sign_with_implicit() {
         let payload =
             b"{\"data\":\"this is a signed message\",\"exp\":\"2022-01-01T00:00:00+00:00\"}";
         let footer = b"{\"kid\":\"zVhMiPBP9fRf2snEcT7gFTioeA9COcNy9DfgL1W60haN\"}";
         let implicit = b"{\"test-vector\":\"4-S-3\"}";
+        let expected = "v4.public.eyJkYXRhIjoidGhpcyBpcyBhIHNpZ25lZCBtZXNzYWdlIiwiZXhwIjoiMjAyMi0wMS0wMVQwMDowMDowMCswMDowMCJ9NPWciuD3d0o5eXJXG5pJy-DiVEoyPYWs1YSTwWHNJq6DZD3je5gf-0M4JR9ipdUSJbIovzmBECeaWmaqcaP0DQ.eyJraWQiOiJ6VmhNaVBCUDlmUmYyc25FY1Q3Z0ZUaW9lQTlDT2NOeTlEZmdMMVc2MGhhTiJ9";
 
-        let one = signing_key.sign_message(payload, None, None);
-        assert_eq!(one, "v4.public.eyJkYXRhIjoidGhpcyBpcyBhIHNpZ25lZCBtZXNzYWdlIiwiZXhwIjoiMjAyMi0wMS0wMVQwMDowMDowMCswMDowMCJ9bg_XBBzds8lTZShVlwwKSgeKpLT3yukTw6JUz3W4h_ExsQV-P0V54zemZDcAxFaSeef1QlXEFtkqxT1ciiQEDA");
+        let key = SigningKey::from_pem(KEY_PEM).expect("decode pem");
 
-        let two = signing_key.sign_message(payload, Some(footer), None);
-        assert_eq!(two, "v4.public.eyJkYXRhIjoidGhpcyBpcyBhIHNpZ25lZCBtZXNzYWdlIiwiZXhwIjoiMjAyMi0wMS0wMVQwMDowMDowMCswMDowMCJ9v3Jt8mx_TdM2ceTGoqwrh4yDFn0XsHvvV_D0DtwQxVrJEBMl0F2caAdgnpKlt4p7xBnx1HcO-SPo8FPp214HDw.eyJraWQiOiJ6VmhNaVBCUDlmUmYyc25FY1Q3Z0ZUaW9lQTlDT2NOeTlEZmdMMVc2MGhhTiJ9");
-
-        let three = signing_key.sign_message(payload, Some(footer), Some(implicit));
-        assert_eq!(three, "v4.public.eyJkYXRhIjoidGhpcyBpcyBhIHNpZ25lZCBtZXNzYWdlIiwiZXhwIjoiMjAyMi0wMS0wMVQwMDowMDowMCswMDowMCJ9NPWciuD3d0o5eXJXG5pJy-DiVEoyPYWs1YSTwWHNJq6DZD3je5gf-0M4JR9ipdUSJbIovzmBECeaWmaqcaP0DQ.eyJraWQiOiJ6VmhNaVBCUDlmUmYyc25FY1Q3Z0ZUaW9lQTlDT2NOeTlEZmdMMVc2MGhhTiJ9");
+        let token = key.sign_message(payload, Some(footer), Some(implicit));
+        assert_eq!(token, expected);
     }
 }
